Refactoring: moved classes and functions into separate files
